Hcap 849 notification on withdrawal (#508)
* WIP

* Added pending acknowledgment status for post hire withdrawals. Work in progress still.

* Ran formatter, working on acknowledgement section

* WIP

* Workflow nearly complete

* Completed workflow, added tosts, fixed issue where the filtered status list was getting infinitely pushed

* Ran formatter, cleared console.logs

* Fixed hide acknowledgement button

* Fixed 849 by adding a status column to show when a user is hired but withdrawn
Fixed 722 by creating a new route. This route handles the case where a HA, upon finding a participant to withdraw from the site details page, can withdraw them and have the employer who hired them notified. This notification takes the form of a 'pending withdrawal' status. This status can made nolong current by acknowledging it as an employer/HA

* Fixed tests by adding the status to the expected columns object.

* Wrapped fetch participants in a useCallback function

* Adding status column to HA column list

* addressing PR comments

* PR polish

When you archive a candidate as an HA, you will not generate a pending withdrawal status if you archive your own candidate.
Fixed withdraw section in the archiveParticipantBySite. Removed siteId from the SiteParticipantTable because it's easily confused with ID.

Co-authored-by: David Bayly <davidbayly@freshworks.io>